posats doubles on han de anara doubles

+++ src/com/company/Producte.java
@@ -22,7 +22,7 @@ public class Producte extends AbstractProduct {
     }
 
     @Override
-    public int ObtenirPreu() {
+    public double ObtenirPreu() {
         return 0;
     }
